[wip] Use npm packages instead of CDN (to be deprecated)
Prepares the codebase for async code splitting (dynamic imports).
By using dynamic imports we'll be able to asynchronously load
chunks only when they're needed. Coupled with ReactLoadable[1],
it should be easy to do this declaratively.

It would still be great if, in production, we had a list of packages
loaded from a CDN (mainly: jQuery, and lodash), and remove them from
the JavaScript bundle.

1. https://github.com/jamiebuilds/react-loadable